abstract getting config from file
